requirements.txt
flask==1.1.1
psycopg2-binary==2.9.7
uWSGI==2.0.21
jinja2==3.0.1
Werkzeug==2.0.3
itsdangerous==2.0.1